move some logs stuff to grafana/ui
