+++ packages/grafana-ui/src/types/index.ts
@@ -7,4 +7,5 @@ export * from './theme';
 export * from './graph';
 export * from './threshold';
 export * from './input';
+export * from './logs';
 export * from './displayValue';
