+++ packages/grafana-ui/src/types/logs.ts
@@ -1,0 +1,21 @@
+/**
+ * Mapping of log level abbreviation to canonical log level.
+ * Supported levels are reduce to limit color variation.
+ */
+export enum LogLevel {
+  emerg = 'critical',
+  alert = 'critical',
+  crit = 'critical',
+  critical = 'critical',
+  warn = 'warning',
+  warning = 'warning',
+  err = 'error',
+  eror = 'error',
+  error = 'error',
+  info = 'info',
+  notice = 'info',
+  dbug = 'debug',
+  debug = 'debug',
+  trace = 'trace',
+  unknown = 'unknown',
+}
