+++ public/app/core/logs_model.ts
@@ -1,29 +1,7 @@
 import _ from 'lodash';
 
-import { colors, TimeSeries } from '@grafana/ui';
+import { colors, TimeSeries, Labels, LogLevel } from '@grafana/ui';
 import { getThemeColor } from 'app/core/utils/colors';
-
-/**
- * Mapping of log level abbreviation to canonical log level.
- * Supported levels are reduce to limit color variation.
- */
-export enum LogLevel {
-  emerg = 'critical',
-  alert = 'critical',
-  crit = 'critical',
-  critical = 'critical',
-  warn = 'warning',
-  warning = 'warning',
-  err = 'error',
-  eror = 'error',
-  error = 'error',
-  info = 'info',
-  notice = 'info',
-  dbug = 'debug',
-  debug = 'debug',
-  trace = 'trace',
-  unknown = 'unknown',
-}
 
 export const LogLevelColor = {
   [LogLevel.critical]: colors[7],
@@ -46,7 +24,7 @@ export interface LogRowModel {
   entry: string;
   hasAnsi: boolean;
   key: string; // timestamp + labels
-  labels: LogsStreamLabels;
+  labels: Labels;
   logLevel: LogLevel;
   raw: string;
   searchWords?: string[];
@@ -54,7 +32,7 @@ export interface LogRowModel {
   timeFromNow: string;
   timeEpochMs: number;
   timeLocal: string;
-  uniqueLabels?: LogsStreamLabels;
+  uniqueLabels?: Labels;
 }
 
 export interface LogLabelStatsModel {
@@ -72,7 +50,7 @@ export enum LogsMetaKind {
 
 export interface LogsMetaItem {
   label: string;
-  value: string | number | LogsStreamLabels;
+  value: string | number | Labels;
   kind: LogsMetaKind;
 }
 
@@ -88,8 +66,8 @@ export interface LogsStream {
   labels: string;
   entries: LogsStreamEntry[];
   search?: string;
-  parsedLabels?: LogsStreamLabels;
-  uniqueLabels?: LogsStreamLabels;
+  parsedLabels?: Labels;
+  uniqueLabels?: Labels;
 }
 
 export interface LogsStreamEntry {
@@ -97,10 +75,6 @@ export interface LogsStreamEntry {
   ts: string;
   // Legacy, was renamed to ts
   timestamp?: string;
-}
-
-export interface LogsStreamLabels {
-  [key: string]: string;
 }
 
 export enum LogsDedupDescription {
